+++ FarmSimulator/src/items/GrowthHormone.java
@@ -10,7 +10,7 @@ public class GrowthHormone extends Food {
 	 * Set the attribute values for growth hormone
 	 */
 	public GrowthHormone() {
-		setHealthiness((float) 1.6);
+		setHealthiness((double) 1.6);
 		setItemName("Growth Hormone");
 		setItemPrice(30);
 		setItemUse("increase an animals healthiness");
